refactor(opentitanlib): execute Phase 3 minor folder reorganization
Physically relocate files to resolve overlapping directory boundaries, achieving 100% on-disk physical target congruence (Phase 3).

Specifically:
- Resolve transport directory overlap (1 file move):
  - Move `src/transport/mod.rs` -> `src/transport.rs` (traits).
  - *Result:* The directory tree `src/transport/` now exclusively and perfectly contains concrete interface driver submodules belonging to the `opentitanlib_transports` target.
  - Update `opentitanlib_core` target in BUILD to point to `src/transport.rs`.
- Resolve proxy directory overlap (6 file moves, 1 creation, 1 removal):
  - Split overlapping `src/proxy/` directory into two congruent directories:
    - Create `src/proxy_protocol/` folder containing the client-server shared serialization modules.
    - Move `src/proxy/errors.rs` -> `src/proxy_protocol/errors.rs`.
    - Move `src/proxy/protocol.rs` -> `src/proxy_protocol/protocol.rs`.
    - Create entry point `src/proxy_protocol/mod.rs` for `opentitanlib_proxy_protocol` target.
    - Delete obsolete flat entry point `src/proxy_protocol_lib.rs`.
    - Rename remaining server submodules directory `src/proxy/` -> `src/proxy_server/`, which now exclusively maps to `opentitanlib_protocols` target.
  - Update protocols entry point (`protocols_lib.rs`) and aggregate `src/lib.rs` re-exports to match.
  - Update target-relative import (`crate::proxy::...` -> `crate::proxy_server::...`) inside `proxy_server/handler.rs`.
  - Update both `opentitanlib_proxy_protocol` and `opentitanlib_protocols` target definitions inside the `BUILD` file.
- Verified: aggregate tests target and downstream tool `opentitantool` compiled and passed successfully under Bazel with zero other files changes!

TAG=agy
CONV=b1350f87-77e1-4940-8d1e-55a68667743e
