+++ sw/host/opentitanlib/BUILD
@@ -103,7 +103,7 @@ rust_library(
         "src/io/uart.rs",
         "src/io/usb/desc.rs",
         "src/io/usb/mod.rs",
-        "src/transport/mod.rs",
+        "src/transport.rs",
         "src/uart/console.rs",
         "src/uart/mod.rs",
     ],
@@ -311,11 +311,11 @@ rust_library(
 rust_library(
     name = "opentitanlib_proxy_protocol",
     srcs = [
-        "src/proxy_protocol_lib.rs",
-        "src/proxy/errors.rs",
-        "src/proxy/protocol.rs",
+        "src/proxy_protocol/errors.rs",
+        "src/proxy_protocol/mod.rs",
+        "src/proxy_protocol/protocol.rs",
     ],
-    crate_root = "src/proxy_protocol_lib.rs",
+    crate_root = "src/proxy_protocol/mod.rs",
     deps = [
         ":opentitanlib_core",
         "@crate_index//:anyhow",
@@ -449,10 +449,10 @@ rust_library(
         "src/otp/lc_state.rs",
         "src/otp/mod.rs",
         "src/otp/otp_img.rs",
-        "src/proxy/handler.rs",
-        "src/proxy/mod.rs",
-        "src/proxy/nonblocking_uart.rs",
-        "src/proxy/socket_server.rs",
+        "src/proxy_server/handler.rs",
+        "src/proxy_server/mod.rs",
+        "src/proxy_server/nonblocking_uart.rs",
+        "src/proxy_server/socket_server.rs",
         "src/rescue/dfu.rs",
         "src/rescue/mod.rs",
         "src/rescue/serial.rs",
