+++ sw/host/opentitanlib/src/proxy_protocol/errors.rs
