+++ sw/host/opentitanlib/src/proxy_server/socket_server.rs
